+++ integrations/azure/azure_integration/ocean.py
@@ -6,27 +6,42 @@ import fastapi
 from loguru import logger
 from starlette import responses
 
-from azure_integration.overrides import AzurePortAppConfig, AzureResourceConfig
+from azure_integration.iterators import (
+    resource_group_iterator,
+    resource_base_kind_iterator,
+    resource_extention_kind_iterator,
+)
+from azure_integration.overrides import (
+    AzurePortAppConfig,
+    AzureSpecificKindSelector,
+    AzureCloudResourceSelector,
+)
 from port_ocean.context.event import event
 from port_ocean.context.ocean import ocean
 from port_ocean.core.models import Entity
 from port_ocean.core.ocean_types import ASYNC_GENERATOR_RESYNC_TYPE
 from azure.identity.aio import DefaultAzureCredential
 from azure.core.exceptions import ResourceNotFoundError
-from azure.mgmt.resource.resources.v2022_09_01.aio import ResourceManagementClient
-
+from azure.mgmt.subscription.aio import SubscriptionClient
 
 from azure_integration.utils import (
     ResourceKindsWithSpecialHandling,
     resource_client_context,
-    get_integration_subscription_id,
     resolve_resource_type_from_resource_uri,
     batch_resources_iterator,
     is_sub_resource,
-    get_resource_kind_by_level,
     get_current_resource_config,
+    stream_async_iterators_tasks,
+    get_resource_configs_with_resource_kind,
 )
-from azure_integration.azure_patch import list_resources
+
+
+def _resolve_resync_method_for_resource(
+    kind: str,
+) -> typing.Callable[..., ASYNC_GENERATOR_RESYNC_TYPE]:
+    if is_sub_resource(kind):
+        return resync_extension_resources
+    return resync_base_resources
 
 
 @ocean.on_resync()
@@ -34,11 +49,13 @@ async def resync_resources(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     if kind in iter(ResourceKindsWithSpecialHandling):
         logger.info("Kind already has a specific handling, skipping", kind=kind)
         return
-    if is_sub_resource(kind):
-        iterator_resync_method = resync_extension_resources
-    else:
-        iterator_resync_method = resync_base_resources
-    async for resources_batch in iterator_resync_method(kind):
+    resource_selector = typing.cast(
+        AzureSpecificKindSelector, get_current_resource_config().selector
+    )
+    iterator_resync_method = _resolve_resync_method_for_resource(kind)
+    async for resources_batch in iterator_resync_method(
+        kind, resource_selector.api_version
+    ):
         yield resources_batch
 
 
@@ -47,29 +64,76 @@ async def resync_resource_groups(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     """
     Re-syncs resource groups, this is done separately because the resource groups api is different from the other apis
     """
+    resource_selector = typing.cast(
+        AzureSpecificKindSelector, get_current_resource_config().selector
+    )
     async with DefaultAzureCredential() as credential:
-        async with ResourceManagementClient(
-            credential=credential, subscription_id=get_integration_subscription_id()
-        ) as client:
-            async for resource_groups_batch in batch_resources_iterator(
-                client.resource_groups.list,
-                api_version=get_current_resource_config().selector.api_version,
+        async with SubscriptionClient(credential=credential) as subscription_client:
+            async for subscriptions_batch in batch_resources_iterator(
+                subscription_client.subscriptions.list,
             ):
-                yield resource_groups_batch
+                tasks = [
+                    resource_group_iterator(
+                        credential=credential,
+                        subscription_id=subscription["subscription_id"],
+                        api_version=resource_selector.api_version,
+                    )
+                    for subscription in subscriptions_batch
+                ]
+                async for resource_groups_batch in stream_async_iterators_tasks(tasks):
+                    yield resource_groups_batch
 
 
-async def resync_base_resources(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
-    async with resource_client_context() as client:
-        async for resources_batch in batch_resources_iterator(
-            list_resources,
-            resources_client=client,
-            resource_type=kind,
-            api_version=get_current_resource_config().selector.api_version,
+@ocean.on_resync(kind=ResourceKindsWithSpecialHandling.SUBSCRIPTION)
+async def resync_subscriptions(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    """
+    Re-syncs subscriptions, this is done separately because the subscriptions api is different from the other apis
+    """
+    async with DefaultAzureCredential() as credential:
+        async with SubscriptionClient(credential=credential) as client:
+            async for subscriptions_batch in batch_resources_iterator(
+                client.subscriptions.list,
+            ):
+                yield subscriptions_batch
+
+
+@ocean.on_resync(kind=ResourceKindsWithSpecialHandling.CLOUD_RESOURCE)
+async def resync_cloud_resources(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    resource_kinds = typing.cast(
+        AzureCloudResourceSelector, get_current_resource_config().selector
+    ).resource_kinds
+    for resource_kind, resource_api_version in resource_kinds.items():
+        iterator_resync_method = _resolve_resync_method_for_resource(resource_kind)
+        async for resources_batch in iterator_resync_method(
+            resource_kind, resource_api_version
         ):
             yield resources_batch
 
 
-async def resync_extension_resources(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
+async def resync_base_resources(
+    kind: str, api_version: str
+) -> ASYNC_GENERATOR_RESYNC_TYPE:
+    async with DefaultAzureCredential() as credential:
+        async with SubscriptionClient(credential=credential) as subscription_client:
+            async for subscriptions_batch in batch_resources_iterator(
+                subscription_client.subscriptions.list,
+            ):
+                tasks = [
+                    resource_base_kind_iterator(
+                        credential=credential,
+                        subscription_id=subscription["subscription_id"],
+                        resource_kind=kind,
+                        api_version=api_version,
+                    )
+                    for subscription in subscriptions_batch
+                ]
+                async for resources_batch in stream_async_iterators_tasks(tasks):
+                    yield resources_batch
+
+
+async def resync_extension_resources(
+    kind: str, api_version: str
+) -> ASYNC_GENERATOR_RESYNC_TYPE:
     """
     Re-syncs sub resources
 
@@ -94,82 +158,26 @@ async def resync_extension_resources(kind: str) -> ASYNC_GENERATOR_RESYNC_TYPE:
     having the base resource ids
 
     :param kind: Resource kind
+    :param api_version: The api version to use when querying the resources
     :return: Async generator of extension resources
     """
     with logger.contextualize(resource_kind=kind):
-        async with resource_client_context() as client:
-            base_resource_kind, _ = get_resource_kind_by_level(kind, 0)
-            api_version = get_current_resource_config().selector.api_version
-            async for resource in list_resources(
-                client,
-                api_version=api_version,
-                resource_type=base_resource_kind,
-            ):
-                async for resource_batch in loop_over_extension_resource_kind(
-                    client=client,
-                    full_resource_kind=kind,
-                    kind_level=1,
-                    resource_id=resource.id,
-                    api_version=api_version,
+        async with DefaultAzureCredential() as credential:
+            async with SubscriptionClient(credential=credential) as subscription_client:
+                async for subscriptions_batch in batch_resources_iterator(
+                    subscription_client.subscriptions.list,
                 ):
-                    yield resource_batch
-
-
-async def loop_over_extension_resource_kind(
-    client: ResourceManagementClient,
-    full_resource_kind: str,
-    kind_level: int,
-    resource_id: str,
-    api_version: str,
-) -> typing.AsyncGenerator[typing.List[dict[str, typing.Any]], None]:
-    """
-    Loops over a extension resource kind and yields a batch of resources
-
-    This method is called recursively until it reaches the last level of the resource kind, and then it yields a batch
-    of resources from the last level.
-
-    :param client: The resource client
-    :param full_resource_kind: Full resource kind (Microsoft.Storage/storageAccounts/blobServices/containers)
-    :param kind_level: The level of the resource kind
-        (0 for Microsoft.Storage/storageAccounts, 1 for Microsoft.Storage/storageAccounts/blobServices, etc)
-    :param resource_id: The resource id of the parent resource, used to query the extension resources of the parent resource
-        (/subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/Microsoft.Storage/storageAccounts/{storageAccountName})
-    :param api_version: The api version to use when querying the resources
-    """
-    current_resource_kind, is_last_level = get_resource_kind_by_level(
-        full_resource_kind, kind_level
-    )
-    current_resource_kind_suffix = current_resource_kind.split("/")[-1]
-    list_resource_url = f"{resource_id}/{current_resource_kind_suffix}"
-    logger.debug(
-        "Looping over resource kind",
-        resource_kind=current_resource_kind,
-        kind_level=kind_level,
-        parent_resource_id=resource_id,
-        api_version=api_version,
-    )
-    if is_last_level:
-        async for resource_batch in batch_resources_iterator(
-            list_resources,
-            resources_client=client,
-            api_version=api_version,
-            resource_url=list_resource_url,
-        ):
-            yield resource_batch
-    else:
-        async for resource in list_resources(
-            client,
-            api_version=api_version,
-            resource_url=list_resource_url,
-        ):
-            async for resource_batch in loop_over_extension_resource_kind(
-                client=client,
-                full_resource_kind=full_resource_kind,
-                kind_level=kind_level + 1,
-                resource_id=resource.id,
-                api_version=api_version,
-            ):
-                yield resource_batch
+                    tasks = [
+                        resource_extention_kind_iterator(
+                            credential=credential,
+                            subscription_id=subscription["subscription_id"],
+                            resource_kind=kind,
+                            api_version=api_version,
+                        )
+                        for subscription in subscriptions_batch
+                    ]
+                    async for resources_batch in stream_async_iterators_tasks(tasks):
+                        yield resources_batch
 
 
 @ocean.router.post("/events")
@@ -181,14 +189,16 @@ async def handle_events(cloud_event: CloudEvent) -> fastapi.Response:
     https://learn.microsoft.com/en-us/azure/event-grid/cloud-event-schema
     """
     cloud_event_data: dict[str, typing.Any] = cloud_event.data  # type: ignore
+    subscription_id = cloud_event_data["subscriptionId"]
+    resource_uri = cloud_event_data["resourceUri"]
     logger.info(
-        "Received azure cloud event",
+        f"Received event {cloud_event.id} of type {cloud_event.type} with operation {cloud_event_data['operationName']} for resource {resource_uri}",
         event_id=cloud_event.id,
         event_type=cloud_event.type,
         resource_provider=cloud_event_data["resourceProvider"],
         operation_name=cloud_event_data["operationName"],
+        subscription_id=subscription_id,
     )
-    resource_uri = cloud_event_data["resourceUri"]
     resource_type = resolve_resource_type_from_resource_uri(
         resource_uri=resource_uri,
     )
@@ -199,48 +209,49 @@ async def handle_events(cloud_event: CloudEvent) -> fastapi.Response:
         )
         return fastapi.Response(status_code=http.HTTPStatus.NOT_FOUND)
 
-    matching_resource_configs: typing.List[AzureResourceConfig] = [
-        resource
-        for resource in typing.cast(AzurePortAppConfig, event.port_app_config).resources
-        if resource.kind == resource_type
-    ]
+    matching_resource_configs = get_resource_configs_with_resource_kind(
+        resource_kind=resource_type,
+        resource_configs=typing.cast(
+            AzurePortAppConfig, event.port_app_config
+        ).resources,
+    )
     if not matching_resource_configs:
-        logger.debug(
+        logger.info(
             "Resource type not found in port app config, update port app config to include the resource type",
             resource_type=resource_type,
         )
         return fastapi.Response(status_code=http.HTTPStatus.NOT_FOUND)
 
-    async with resource_client_context() as client:
+    async with resource_client_context(subscription_id) as client:
         for resource_config in matching_resource_configs:
+            if isinstance(resource_config.selector, AzureSpecificKindSelector):
+                api_version = resource_config.selector.api_version
+            else:
+                api_version = resource_config.selector.resource_kinds[resource_type]
             blueprint = resource_config.port.entity.mappings.blueprint.strip('"')
-            logger.debug(
-                "Querying full resource",
-                id=resource_uri,
-                kind=resource_type,
-                api_version=resource_config.selector.api_version,
-                blueprint=blueprint,
+            logger.info(
+                f"Querying full resource details for resource {resource_uri}, api version {api_version} for kind {resource_type} in subscription {subscription_id}"
             )
             try:
                 resource = await client.resources.get_by_id(
                     resource_id=resource_uri,
-                    api_version=resource_config.selector.api_version,
+                    api_version=api_version,
                 )
-                await ocean.register_raw(resource_type, [dict(resource.as_dict())])
+                await ocean.register_raw(
+                    resource_config.kind, [dict(resource.as_dict())]
+                )
             except ResourceNotFoundError:
                 logger.info(
                     "Resource not found in azure, unregistering from port",
                     id=resource_uri,
-                    kind=resource_type,
-                    api_version=resource_config.selector.api_version,
+                    kind=resource_config.kind,
+                    api_version=api_version,
                     blueprint=blueprint,
                 )
                 await ocean.unregister(
                     [
                         Entity(
-                            blueprint=resource_config.port.entity.mappings.blueprint.strip(
-                                '"'
-                            ),
+                            blueprint=blueprint,
                             identifier=resource_uri,
                         )
                     ]